Game search component template
