--- src/components/tier-list/styles/searchGame.style.css
@@ -1,3 +1,0 @@
-.findGameBtn{
-    border-radius: 0 5px 5px 0;
-}
